example/pyproject.toml
[project]
name = "fastapi-getpaid-example"
version = "0.0.0"
description = "Example application for fastapi-getpaid with fake payment gateway"
requires-python = ">=3.12"
dependencies = [
    "fastapi-getpaid[sqlalchemy]",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

